Selecting a shared sub-segment highlights sharing resolved topologies.
When the user clicks on a shared sub-segment the resolved topologies
will get added to the selected features list. The user can still click
on the (topological section) feature of the shared sub-segment
(provided visibility of topological sections is enabled), in which case
the full topological section feature will get highlighted
(including the dangling bits at the ends, if it's a line, and not just
the sub-segments that actually contribute to the resolved topologies).